+++ precompiles/wasmd/integration_test.go
@@ -21,9 +21,9 @@ import (
 	"github.com/cosmos/evm/x/vm/statedb"
 	evmtypes "github.com/cosmos/evm/x/vm/types"
 
-	app "github.com/kiichain/kiichain/v5/app"
-	helpers "github.com/kiichain/kiichain/v5/app/helpers"
-	wasmdprecompile "github.com/kiichain/kiichain/v5/precompiles/wasmd"
+	app "github.com/kiichain/kiichain/v6/app"
+	helpers "github.com/kiichain/kiichain/v6/app/helpers"
+	wasmdprecompile "github.com/kiichain/kiichain/v6/precompiles/wasmd"
 )
 
 // CounterWasmCode is the bytecode of the counter smart contract