+++ tests/e2e/evm_util.go
@@ -16,7 +16,7 @@ import (
 	geth "github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/ethclient"
 
-	kiichain "github.com/kiichain/kiichain/v5/app"
+	kiichain "github.com/kiichain/kiichain/v6/app"
 )
 
 // EVMAccount stores an address and a key used for EVM interaction
@@ -96,6 +96,53 @@ func EVMTransaction(
 		return geth.Receipt{}, fmt.Errorf("receipt status is 0")
 	}
 	return *receipt, nil
+}
+
+// EVMSendWithNonce builds up an evm_sendTransaction with a specific nonce and returns its tx
+func EVMSendWithNonce(
+	client *ethclient.Client,
+	privateKey *ecdsa.PrivateKey,
+	fromAddress common.Address,
+	toAddress common.Address,
+	amount *big.Int,
+	contractBinary []byte,
+	nonce uint64,
+) (*geth.Transaction, error) {
+	// Get suggested gas price
+	gasPrice, err := client.SuggestGasPrice(context.Background())
+	if err != nil {
+		return nil, fmt.Errorf("failed to get gas price: %w", err)
+	}
+
+	// Get chain ID
+	chainID, err := client.NetworkID(context.Background())
+	if err != nil {
+		return nil, fmt.Errorf("failed to get chain ID: %w", err)
+	}
+
+	// Create the transaction
+	tx := geth.NewTransaction(
+		nonce,
+		toAddress,
+		amount,
+		1500000, // gas limit
+		gasPrice,
+		contractBinary, // contract bytes
+	)
+
+	// Sign the transaction
+	signedTx, err := geth.SignTx(tx, geth.NewEIP155Signer(chainID), privateKey)
+	if err != nil {
+		return nil, fmt.Errorf("failed to sign transaction: %w", err)
+	}
+
+	// Send the transaction
+	err = client.SendTransaction(context.Background(), signedTx)
+	if err != nil {
+		return nil, fmt.Errorf("failed to send transaction: %w", err)
+	}
+
+	return signedTx, nil
 }
 
 // checkTransactionByHash returns a receipt from a given tx hash