+++ tests/integration/precompiles/p256/precompile_p256_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/cosmos/evm/tests/integration/precompiles/p256"
 
-	"github.com/kiichain/kiichain/v5/tests/integration"
+	"github.com/kiichain/kiichain/v6/tests/integration"
 )
 
 func TestP256PrecompileTestSuite(t *testing.T) {