+++ x/feeabstraction/ante/evm/mono_decorator.go
@@ -30,7 +30,7 @@ import (
 	evmkeeper "github.com/cosmos/evm/x/vm/keeper"
 	evmtypes "github.com/cosmos/evm/x/vm/types"
 
-	antetypes "github.com/kiichain/kiichain/v5/ante/types"
+	antetypes "github.com/kiichain/kiichain/v6/ante/types"
 )
 
 const AcceptedTxType = 0 |