+++ x/oracle/abci_test.go
@@ -8,9 +8,9 @@ import (
 	"cosmossdk.io/collections"
 	"cosmossdk.io/math"
 
-	"github.com/kiichain/kiichain/v5/x/oracle/keeper"
-	"github.com/kiichain/kiichain/v5/x/oracle/types"
-	"github.com/kiichain/kiichain/v5/x/oracle/utils"
+	"github.com/kiichain/kiichain/v6/x/oracle/keeper"
+	"github.com/kiichain/kiichain/v6/x/oracle/types"
+	"github.com/kiichain/kiichain/v6/x/oracle/utils"
 )
 
 /* SetUp conditions: