+++ x/rewards/client/cli/tx.go
@@ -10,7 +10,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/client/tx"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	"github.com/kiichain/kiichain/v5/x/rewards/types"
+	"github.com/kiichain/kiichain/v6/x/rewards/types"
 )
 
 // GetTxCmd returns the transaction commands for this module