+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## Unreleased
+## v6.0.0 - 2025-11-13
 
 ## Added
 
@@ -8,11 +8,13 @@
 - Added logs and telemetry for reentrance detection in wasmd precompile
 - Add further validations to Wasm Oracle query bindings
 - Moves Wasmd reentrance lock to the core of the wasmd contract to avoid reentrance attacks on queries and instantiations
+- Added EVM mempool
 
 ## Removed
 
 - Removed IBC precompile since ICS20 precompile also handles IBC transfers.
 - Removed fallback native price param on the fee abstraction module. Instead of using hardcoded price, the module disables itself when price is lacking.
+- Removed previous upgrades from main branch.
 
 ## DEPENDENCIES
 
@@ -25,6 +27,7 @@
 - Fix Oracle module ante decorator to allow first vote and install oracle ante handlers
 - Fix IBC validation of negative numbers happens a bit early [#144](https://github.com/KiiChain/kiichain/issues/144)
 - Fix incorrect error passing on tokenfactory wasmbinding
+- Fix blocked address checked after minting in wasmbinding tokenfactory 
 - Fix account balance information being stale on fee abstraction's cosmos' ante
 
 ### Documentation